fixed warning
git-svn-id: http://svn.seqan.de/seqan/trunk@7991 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/projects/benchmarks/read_mappers/apps/test_find_myers_ukkonen_reads.cpp
@@ -241,7 +241,7 @@ SEQAN_DEFINE_TEST(test_find_myers_ukkonen_set_end_position2) {
     DnaString hstck = "ATTTCGGTCATCAAATAATCATTTATTTTGCCACAACATAAAAATAATTGTCTGAATATGGAATTGTCTGAACCTCACTGAGCTCGTAATAAAATTTCCA";
     Finder<DnaString> finder(hstck);
     DnaString ndl = "CACAACATAAAAAATAATTGTCTGAATATGGAATGT";
-    Pattern<DnaString, Myers<FindInfix> > pattern(ndl, -length(ndl));
+    Pattern<DnaString, Myers<FindInfix> > pattern(ndl, -static_cast<int>(length(ndl)));
 
     bool ret = setEndPosition(finder, pattern, 66u);
     SEQAN_ASSERT_TRUE(ret);
